Update transformer.py
For now, fix the model we need in the AutoModel. Later, we can try to adapt the actual saved model as this is the issue. It is referring to the bert base model, but the vocab size changed sinced we trained our model, causing an issue.

Also we clear kwargs as those cause an error as well.